# GridExpand/docs/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Jobs, results and scenario files of GridExpand (load allocation, urbs optimisation, power flow, grid expansion) for the GridPlanner UI. API version 1.",
    "title": "GridExpand API",
    "version": "api-1",
    "x-api-version": 1
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/results/analyses": {
      "get": {
        "description": "``expansion_analysis_run`` rows with totals (cost, cables and transformers to reinforce).\n\nWith ``plz`` (and/or ``pylovo_version_id``) the totals cover only those grids and\nanalyses without such grids are left out.",
        "operationId": "analyses_api_results_analyses_get",
        "parameters": [
          {
            "in": "query",
            "name": "ags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ags"
            }
          },
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 99999,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          },
          {
            "in": "query",
            "name": "pylovo_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[\\w.-]{1,10}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pylovo Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Analyses Api Results Analyses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyses",
        "tags": [
          "results"
        ]
      }
    },
    "/api/results/analyses/{analysis_key}/assets": {
      "get": {
        "description": "Building assets of the analysed power-flow runs: one point per building (EPSG:4326)\nwith PV, battery, heat pump, heating rod, heat storage and EV capacities, plus totals\n(``features=false``: totals only).",
        "operationId": "analysis_assets_api_results_analyses__analysis_key__assets_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_key",
            "required": true,
            "schema": {
              "title": "Analysis Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 99999,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          },
          {
            "in": "query",
            "name": "pylovo_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[\\w.-]{1,10}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pylovo Version Id"
            }
          },
          {
            "in": "query",
            "name": "features",
            "required": false,
            "schema": {
              "default": true,
              "title": "Features",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Analysis Assets Api Results Analyses  Analysis Key  Assets Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis Assets",
        "tags": [
          "results"
        ]
      }
    },
    "/api/results/analyses/{analysis_key}/geojson": {
      "get": {
        "description": "Cables (``action``: none / add_1 / add_2plus) and transformers of one analysis (EPSG:4326).",
        "operationId": "analysis_geojson_api_results_analyses__analysis_key__geojson_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_key",
            "required": true,
            "schema": {
              "title": "Analysis Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 99999,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          },
          {
            "in": "query",
            "name": "pylovo_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[\\w.-]{1,10}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pylovo Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Analysis Geojson Api Results Analyses  Analysis Key  Geojson Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis Geojson",
        "tags": [
          "results"
        ]
      }
    },
    "/api/results/analyses/{analysis_key}/grids": {
      "get": {
        "description": "Per-grid cost, cables to reinforce and transformer loading of one analysis.",
        "operationId": "analysis_grids_api_results_analyses__analysis_key__grids_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_key",
            "required": true,
            "schema": {
              "title": "Analysis Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 99999,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          },
          {
            "in": "query",
            "name": "pylovo_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[\\w.-]{1,10}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pylovo Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Analysis Grids Api Results Analyses  Analysis Key  Grids Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analysis Grids",
        "tags": [
          "results"
        ]
      }
    },
    "/api/results/powerflow": {
      "get": {
        "description": "``powerflow_summary`` per grid, run (model case) and stage: loading percentiles, voltage p05, hours above limits.",
        "operationId": "powerflow_api_results_powerflow_get",
        "parameters": [
          {
            "in": "query",
            "name": "ags",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ags"
            }
          },
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 99999,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          },
          {
            "in": "query",
            "name": "pylovo_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[\\w.-]{1,10}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pylovo Version Id"
            }
          },
          {
            "in": "query",
            "name": "scenario_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Powerflow Api Results Powerflow Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Powerflow",
        "tags": [
          "results"
        ]
      }
    },
...
  }
}